requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.1
django-crispy-forms==2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-tinymce==3.6.1
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
